reorganise code, add trueEntropyCalc global flag
