+++ src/util/globals.js
@@ -137,3 +137,5 @@ export const notificationDuration = 5000;
 
 /* server init stuff */
 export const charonAPIAddress = "/charon?"
+
+export const trueEntropyCalc = false;
